requirements.txt
selenium>=4.15.0
webdriver-manager>=4.0.0
requests>=2.31.0
schedule>=1.2.0
qdrant-client>=1.7.0
sentence-transformers>=2.2.0
torch>=2.2.0
transformers>=4.35.0
numpy>=1.24.0
python-dotenv>=1.0.0
streamlit>=1.28.0
openai>=1.0.0 